Generalize ChangeTimer and use it in Mathjax editor
