+++ ts/editor/input-handlers.ts
@@ -8,7 +8,7 @@
 import { nodeIsElement } from "lib/dom";
 import { updateActiveButtons } from "./toolbar";
 import { EditingArea } from "./editing-area";
-import { triggerChangeTimer } from "./change-timer";
+import { triggerChangeTimer } from "./saving";
 import { registerShortcut } from "lib/shortcuts";
 
 export function onInput(event: Event): void {
